+++ extension/util/entprops.h
@@ -87,7 +87,7 @@ public:
 	bool SetEntPropEnt(int entity, PropType proptype, const char *prop, int other, int element = 0);
 	bool GetEntPropVector(int entity, PropType proptype, const char *prop, Vector &result, int element = 0);
 	bool SetEntPropVector(int entity, PropType proptype, const char *prop, Vector value, int element = 0);
-	bool GetEntPropString(int entity, PropType proptype, const char* prop, char* result, int maxlen, int& len, int element = 0);
+	bool GetEntPropString(int entity, PropType proptype, const char* prop, char* result, int maxlen, size_t& len, int element = 0);
 	bool SetEntPropString(int entity, PropType proptype, const char *prop, char *value, int element = 0);
 	bool GetEntData(int entity, int offset, int &result, int size = 4);
 	bool SetEntData(int entity, int offset, int value, int size = 4, bool changeState = false);
@@ -97,13 +97,13 @@ public:
 	bool SetEntDataEnt(int entity, int offset, int value, bool changeState = false);
 	bool GetEntDataVector(int entity, int offset, Vector &result);
 	bool SetEntDataVector(int entity, int offset, Vector value, bool changeState = false);
-	bool GetEntDataString(int entity, int offset, int maxlen, char* result, int &len);
+	bool GetEntDataString(int entity, int offset, int maxlen, char* result, size_t &len);
 	bool SetEntDataString(int entity, int offset, char *value, int maxlen, bool changeState = false);
 	bool GameRules_GetProp(const char *prop, int &result, int size = 4, int element = 0);
 	bool GameRules_GetPropFloat(const char *prop, float &result, int element = 0);
 	bool GameRules_GetPropEnt(const char *prop, int &result, int element = 0);
 	bool GameRules_GetPropVector(const char *prop, Vector &result, int element = 0);
-	bool GameRules_GetPropString(const char *prop, char* result, int& len, int maxlen, int element = 0);
+	bool GameRules_GetPropString(const char *prop, char* result, size_t& len, int maxlen, int element = 0);
 	RoundState GameRules_GetRoundState();
 
 private:
